ref(xmpp): validate the sender of client-requirements and visitors IQs.
Handle the client-requirements IQ in ChatRoom, and accept it only from the
focus, like the mute IQs. Accept a visitors IQ only when the MUC service
addressed it to us.
